Add files via upload
Changes made to update code to be php 5 friendly, add a few layers of extra security, and update to mysqli so this engine can be used for many more years to come.  Security updates are only minimal, add your own, or use at your own risk.  All credit goes to renderse7en for creating this engine and releasing it to the community to let more game developers enjoy it.