captch fixes, added to login if login fails
captch fixes, added to login if login fails so that if a username or password is not correct on a second attempt, user must use captcha.  This is to reduce brute forcing password attacks.  Change password has also been removed from login page, a user should already be logged in to intentionally change a password.